womparator/backend/requirements.txt
flask
flask_cors
python-docx
pandas
werkzeug
